+++ src/Componenets/redux/search/Api.js
@@ -1,10 +1,7 @@
-const axios = require("axios");
-export const SEARCH = "search/type/SEARCH";
+import { Search, SearchError } from "./action";
+import axios from 'axios'
 
-export const Search = (payload) => ({
-    type: SEARCH,
-    payload,
-    })
+
 export const getSearchResults = () => (dispatch) => {
   console.log("SEARCN STARTED ------------------ ");
     const options = {
@@ -12,24 +9,16 @@ export const getSearchResults = () => (dispatch) => {
       url: 'https://contextualwebsearch-websearch-v1.p.rapidapi.com/api/Search/WebSearchAPI',
       params: {q: 'taylor swift', pageNumber: '1', pageSize: '10', autoCorrect: 'true'},
       headers: {
-        'X-RapidAPI-Key': 'SIGN-UP-FOR-KEY',
+        'X-RapidAPI-Key': '0a632a57aamsh1f8242879ce4ecfp1a7230jsnfa28d8ec3311',
         'X-RapidAPI-Host': 'contextualwebsearch-websearch-v1.p.rapidapi.com'
       }
     };
-    
+    console.log("REQUESTING.................");
     axios.request(options).then(function (response) {
         console.log("SEARCH RESULT: ",response.data);
         dispatch(Search(response.data));
     }).catch(function (error) {
+        dispatch(SearchError(error))
         console.error(error);
     });
 }
-const searchReducer = (state = [], action) => {
-    switch (action.type) {
-        case SEARCH:
-            return action.payload;
-        default:
-            return state;
-    }
-}
-export default searchReducer;
