+++ src/Componenets/redux/search/action.js
@@ -1,0 +1,11 @@
+import { SEARCH, SEARCHERROR } from "./type";
+
+export const Search = (payload) => ({
+    type: SEARCH,
+    payload,
+    })
+
+export const SearchError = (payload) => ({
+type: SEARCHERROR,
+payload,
+})
