+++ src/Componenets/redux/search/type.js
@@ -1,0 +1,2 @@
+export const SEARCH = "search/type/SEARCH";
+export const SEARCHERROR = "search/type/SEARCHERROR";
