+++ src/Componenets/redux/storeConfig.js
@@ -1,9 +1,15 @@
-import { legacy_createStore as createStore } from "redux";
-import searchReducer from "./search/search";
-import { combineReducers } from "redux";
+import thunk from 'redux-thunk';
+import {
+  combineReducers,
+  legacy_createStore as createStore,
+  applyMiddleware,
+} from 'redux';
+import searchReducer from "./search/reducer";
+
 
 const rootReducer =combineReducers ({
     searchReducer
 })
-const store =  createStore(rootReducer,{});
+const initState = {};
+const store = createStore(rootReducer, initState, applyMiddleware(thunk));
 export default store;
